requirements.txt
pillow>=9.5.0
pyinstaller>=5.13.0
